@@ -12,7 +12,7 @@
     "start": "node .output/server/index.mjs",
     "generate": "nuxi generate",
     "preview": "nuxi preview",
-    "postinstall": "nuxi prepare; node scripts/postinstall.js; husky install",
+    "postinstall": "nuxi prepare; husky install",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "release:patch": "standard-version --release-as patch",
@@ -21,41 +21,43 @@
     "vercel": "vercel --prod"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^0.36.0",
+    "@antfu/eslint-config": "^0.38.4",
     "@iconify-json/carbon": "^1.1.16",
     "@iconify-json/twemoji": "^1.1.11",
-    "@iconify/json": "^2.2.35",
+    "@iconify/json": "^2.2.49",
     "@kevinmarrec/nuxt-pwa": "^0.17.0",
-    "@nuxt-themes/typography": "^0.10.0",
+    "@nuxt-themes/typography": "^0.11.0",
     "@nuxt/content": "^2.5.2",
-    "@nuxt/image-edge": "^1.0.0-28008586.381b906",
+    "@nuxt/devtools": "npm:@nuxt/devtools-edge@latest",
+    "@nuxt/image-edge": "1.0.0-28020728.5df24eb",
     "@nuxtjs/color-mode": "^3.2.0",
     "@pinia/nuxt": "^0.4.7",
     "@playwright/test": "^1.31.2",
     "@unocss/nuxt": "^0.50.4",
-    "@unocss/preset-icons": "^0.50.4",
+    "@unocss/preset-icons": "^0.51.4",
     "@unocss/preset-typography": "^0.50.4",
     "@vueuse/nuxt": "^9.13.0",
     "@windicss/config": "^1.8.10",
     "@windicss/plugin-animations": "^1.0.9",
     "@windicss/plugin-heropatterns": "^0.0.8",
     "@windicss/plugin-utils": "^1.8.10",
-    "eslint": "^8.36.0",
+    "eslint": "8.38.0",
     "husky": "^8.0.3",
     "less": "^4.1.3",
     "lint-staged": "^13.2.0",
     "lite-youtube-embed": "^0.2.0",
-    "nuxt": "^3.3.3",
+    "nuxt": "^3.4.1",
     "nuxt-pwa-public-manifest": "^1.0.0",
-    "nuxt-windicss": "^2.6.0",
+    "nuxt-windicss": "^2.6.1",
     "pinia": "^2.0.33",
-    "sass": "^1.59.3",
+    "sass": "^1.62.0",
     "sitemap": "^7.1.1",
     "standard-version": "^9.5.0",
     "unplugin-icons": "^0.15.3"
   },
   "dependencies": {
     "@nuxtjs/robots": "^3.0.0",
-    "ipx": "^1.0.0-2"
+    "ipx": "1.0.0-2",
+    "untyped": "^1.3.2"
   }
 }
