@@ -1,5 +1,8 @@
-const fs = require("fs");
-const consola = require("consola");
+// const fs = require("fs");
+import * as fs from 'fs';
+
+// const consola = require("consola");
+import * as consola from 'consola';
 
 (function () {
   // Create an .env file if one doesn't exist
